refactor: create OfferForm component
