+++ freex/src/components/MyOffers/MyOffers.jsx
@@ -40,6 +40,7 @@ const MyOffers = () => {
       const data = await getDocs(q);
       const filteredData = data.docs.map((doc) => ({
         ...doc.data(),
+        id: doc.id,
       }));
 
       setUserOffers(filteredData);
